Fixed routing

+++ Selstice.Data/Product.cs
@@ -10,7 +10,8 @@ namespace Selstice.Data
 	{
 		public string Name { get; set; }
 		public string Description { get; set; }
-		public double Price { get; set; }
+		public decimal Price { get; set; }
 		public ICollection<ProductPriceHistory> PriceHistory { get; set; }
 	}
 }
+
